+++ packages/interact/CHANGELOG.md
@@ -15,7 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Support for standalone execution (ie without a `package.json` file)
 * Added pagefind support
 * Added HTML cache while browsing to support pagefind indexing
-* Added a SearchBox (URL preview, command box, ...)
+* Added a [SearchBox](src/resources/components/interact/SearchBox.tsx) (URL preview, command box, ...)
 * Added eslint to check that the node directory does not import module via the `@combostrap` import syntax
 
 ## [0.1.1] - 2026-06-18